requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
discord.py==1.3.2
google==2.0.3
idna==2.9
multidict==4.7.5
python-dotenv==0.12.0
soupsieve==2.0
websockets==8.1
wincertstore==0.2
yarl==1.4.2
